Move exception reporting into `Subscribers::ProcessAction`... (#20)
This lets us fix the "missing status code" issue (#11) closer to where it
occurs. (The "`status` is nil if an exception occurred" issue appears to
be a bug in ActiveSupport::Notifications, since the standard Rails
logging also
[handles this case explicitly](https://github.com/rails/rails/blob/37b373a8d2a1cd132bbde51cd5a3abd4ecee433b/actionpack/lib/action_controller/log_subscriber.rb#L25).)
The fix could live anywhere, but this makes it clearer what we're
working around, so we don't need to bring along this workaround if we
decide to move the logic somewhere else (e.g. a Rack middleware).

It also simplifies the tests, and probably removes a small amount of
overhead due to getting rid of the around_filter.