Add integration tests and test on multiple Rails versions (#21)
This adds integration tests - we set up a minimal Rails app, install and configure the Railtie, fire some requests, and verify the right events were sent.

(This uses Rails' built-in integration testing support (via rspec-rails's request specs), so it bypasses the UI but otherwise exercises most of the stack in a fairly realistic way.)

Prior to this commit the tests were only at the unit level, testing individual pieces of the integration. However, the stuff most likely to break - and therefore that we'd most like test coverage for - is how we hook those pieces into the Rails lifecycle, and how they interact. This is particularly true when we consider supporting multiple Rails versions - where a hook we depend on may not exist - or differently configured Rails apps, or interaction with other 3rd-party Rails plugins. It's also helpful to have these in place if we're reimplementing functionality via a different mechanism (such as #20).

This also asks TravisCI to run each build on multiple Rails versions (as well as multiple Ruby versions) - here's an example build confirming we support Rails 3, 4 and 5.